doc_embedding/requirements.txt
langchain==0.3.19
langchain-huggingface==0.1.2
sentence-transformers==4.0.2
faiss-cpu==1.10.0
langchain_community==0.3.18
beautifulsoup4==4.13.3
readabilipy==0.3.0
accelerate==1.6.0